+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Changed
+- Requires Java 21
+
+### Added
+- CrossConcat function
+- CrossConcatSequence function
+
 ## [1.3.3] - 2024-09-19
 
 ### Fixed